undrill copyAsCurl
