+++ packages/insomnia/src/ui/components/wrapper.tsx
@@ -128,7 +128,6 @@ export type Props = ReturnType<typeof mapStateToProps> & ReturnType<typeof mapDi
   handleSetSidebarFilter: (value: string) => Promise<void>;
   handleSetActiveEnvironment: (environmentId: string | null) => Promise<void>;
   handleDuplicateRequest: Function;
-  handleCopyAsCurl: Function;
   handleSetResponsePreviewMode: Function;
   handleSetResponseFilter: Function;
   handleSetActiveResponse: Function;
